+++ internal/providers/bsl/rules/new_objects.go
@@ -14,20 +14,20 @@ func ObjectDefinitionKey() models.RepositoryKey {
 func ObjectDefinition() *models.Rule {
 	patterns := []*models.Rule{
 		{
-			Match: fmt.Sprintf(`(?i:(%s)\\s+)([_$[:alpha:]][_$[:alnum:]]*\\b)`, bslm.BslNew),
+			Match: fmt.Sprintf(`(?i:(%s)\s+)(%s)`, bslm.BslNew, bslm.Identifier),
 			Captures: map[string]models.Capture{
 				"1": {Name: "keyword.operator.new.bsl"},
 				"2": {Name: "support.class.bsl"},
 			},
 		},
 		{
-			Begin: fmt.Sprintf(`(?i:(%s)\\s+)([_$[:alpha:]][_$[:alnum:]]*\\b)(?:\\s*)(%s)\\s*`, bslm.BslNew, bslm.BslParenOpen),
+			Begin: fmt.Sprintf(`(?i:(%s)\s+)(%s)(?:\s*)(\()\s*`, bslm.BslNew, bslm.Identifier),
 			BeginCaptures: map[string]models.Capture{
 				"1": {Name: "keyword.operator.new.bsl"},
 				"2": {Name: "support.class.bsl"},
 				"3": {Name: "punctuation.bracket.begin.bsl"},
 			},
-			End: fmt.Sprintf(`\\s*(%s)`, bslm.BslParenClose),
+			End: `\s*(\))`,
 			EndCaptures: map[string]models.Capture{
 				"1": {Name: "punctuation.bracket.end.bsl"},
 			},